refactor(web): features構成をReact標準（components/hooks）へ変更
ui/model/lib のFSDセグメント命名はroutes/の既存構成と食い違い、2スライス
規模では過分割となるため、routes/と揃えた構成へ統一する。Public API
（index.ts）と外部参照の集約は維持する。

- ui/ → components/
- model/use-session → hooks/
- model/validation・types、lib/* はスライス直下へ
