+++ application/packages/web/src/client/features/diary/components/read-pagination.tsx
@@ -1,4 +1,4 @@
-import type { ReadPagination } from '@/client/features/diary/model/types'
+import type { ReadPagination } from '@/client/features/diary/types'
 
 interface Props {
   onNextPage: () => void
