+++ application/packages/web/src/client/features/diary/index.ts
@@ -1,7 +1,7 @@
-export { default as DiaryLoginRequired } from './ui/login-required'
-export { default as DiaryPageLayout } from './ui/page-layout'
-export { default as DiaryReadListSection } from './ui/read-list-section'
-export { default as DiaryReadPagination } from './ui/read-pagination'
-export { default as DiarySummarySection } from './ui/summary-section'
-export { getTodayJst, sumSourceSummary } from './lib/daily-summary'
-export type { ReadItem, ReadPagination, Source, Summary } from './model/types'
+export { default as DiaryLoginRequired } from './components/login-required'
+export { default as DiaryPageLayout } from './components/page-layout'
+export { default as DiaryReadListSection } from './components/read-list-section'
+export { default as DiaryReadPagination } from './components/read-pagination'
+export { default as DiarySummarySection } from './components/summary-section'
+export { getTodayJst, sumSourceSummary } from './daily-summary'
+export type { ReadItem, ReadPagination, Source, Summary } from './types'
